requirements.txt
black==23.3.0
blinker==1.6.2
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
ed25519==1.5
Flask==2.3.1
idna==2.8
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
nats-py==2.2.0
nkeys==0.1.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pycairo==1.23.0
PyGObject==3.36.0
python-apt==2.0.1+ubuntu0.20.4.1
python-snappy==0.6.1
requests==2.22.0
requests-unixsocket==0.2.0
six==1.14.0
ssh-import-id==5.10
tomli==2.0.1
typing_extensions==4.5.0
unattended-upgrades==0.1
urllib3==1.25.8
Werkzeug==2.3.2
zipp==3.15.0
